vertebrae-api/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
black==25.11.0
boto3==1.34.34
botocore==1.34.34
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@525208d9d6b36c72500183470fd181848b822468
exceptiongroup==1.3.0
fastapi==0.109.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
fvcore==0.1.5.post20221221
grpcio==1.76.0
h11==0.16.0
httptools==0.7.1
hydra-core==1.3.2
idna==3.11
iopath==0.1.9
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.5.0
portalocker==3.2.0
protobuf==6.33.1
pycocotools==2.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytokens==0.3.0
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
tomli==2.3.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.15.0
urllib3==2.0.7
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
yacs==0.1.8
